backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
groq==0.12.0
python-dotenv==1.0.1